requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
google-genai==0.3.0
pdf2image==1.17.0
pydantic==2.6.1
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
